+++ src/components/App.styled.js
@@ -1,40 +1,40 @@
-import styled from "styled-components";
-import { NavLink } from "react-router-dom";
+// import styled from "styled-components";
+// import { NavLink } from "react-router-dom";
 
-export const Container = styled.div`
-  max-width: 960px;
-  margin: 0 auto;
-  padding: 0 16px;
-`;
+// export const Container = styled.div`
+//   max-width: 960px;
+//   margin: 0 auto;
+//   padding: 0 16px;
+// `;
 
-export const Header = styled.header`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  padding: 8px 0;
-  margin-bottom: 16px;
-  border-bottom: 1px solid black;
+// export const Header = styled.header`
+//   display: flex;
+//   align-items: center;
+//   justify-content: space-between;
+//   gap: 12px;
+//   padding: 8px 0;
+//   margin-bottom: 16px;
+//   border-bottom: 1px solid black;
 
-  > nav {
-    display: flex;
-  }
-`;
+//   > nav {
+//     display: flex;
+//   }
+// `;
 
-export const Logo = styled.p`
-  font-weight: 700;
-  margin: 0;
-`;
+// export const Logo = styled.p`
+//   font-weight: 700;
+//   margin: 0;
+// `;
 
-export const Link = styled(NavLink)`
-  padding: 8px 16px;
-  border-radius: 4px;
-  text-decoration: none;
-  color: black;
-  font-weight: 500;
+// export const Link = styled(NavLink)`
+//   padding: 8px 16px;
+//   border-radius: 4px;
+//   text-decoration: none;
+//   color: black;
+//   font-weight: 500;
 
-  &.active {
-    color: white;
-    background-color: orangered;
-  }
-`;
+//   &.active {
+//     color: white;
+//     background-color: orangered;
+//   }
+// `;
