--- src/components/Header/Header.jsx
@@ -1,8 +1,0 @@
-
-export const Header = () => {
-    return (
-      <>
-        
-      </>
-    );
-  };
